+++ src/main/java/de/fsr/mariokart_backend/match_plan/model/Round.java
@@ -33,6 +33,8 @@ public class Round {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    private int roundNumber;
+
     private LocalDateTime startTime;
 
     private LocalDateTime endTime;
