+++ src/main/java/de/fsr/mariokart_backend/match_plan/model/dto/RoundFromBreakReturnDTO.java
@@ -12,6 +12,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class RoundFromBreakReturnDTO {
     private Long id;
+    private int roundNumber;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private boolean finalGame;
