+++ src/main/java/de/fsr/mariokart_backend/match_plan/model/dto/RoundFromGameReturnDTO.java
@@ -12,6 +12,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class RoundFromGameReturnDTO {
     private Long id;
+    private int roundNumber;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private boolean finalGame;
