+++ src/main/java/de/fsr/mariokart_backend/match_plan/model/dto/RoundReturnDTO.java
@@ -13,6 +13,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class RoundReturnDTO {
     private Long id;
+    private int roundNumber;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private boolean finalGame;
