+++ src/main/java/de/fsr/mariokart_backend/match_plan/repository/RoundRepository.java
@@ -2,6 +2,7 @@ package de.fsr.mariokart_backend.match_plan.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -14,4 +15,5 @@ public interface RoundRepository extends JpaRepository<Round, Long> {
     List<Round> findByStartTimeAfter(LocalDateTime startTime);
     List<Round> findByStartTimeBefore(LocalDateTime startTime);
     List<Round> findByPlayedFalse();
+    Optional<Round> findByRoundNumber(int roundNumber);
 }
